P27_Refactor_RoleModule used bool instead of List

+++ Postieri/Services/RolesService.cs
@@ -14,33 +14,38 @@ namespace Postieri.Services
         {
             return _context.Roles.ToList();
         }
-        public List<Role> AddRole(Role role)
+        public bool AddRole(Role request)
         {
-            var _role = new Role()
+            var role = new Role()
             {
-                Name = role.Name,
-                Description = role.Description,
+                Name = request.Name,
+                Description = request.Description,
             };
-            _context.Roles.Add(_role);
+            if(role == null)
+                return false;
+            _context.Roles.Add(role);
             _context.SaveChanges();
-            return _context.Roles.ToList();
+            return true;
         }
-        public List<Role> UpdateRole(Role request)
+        public bool UpdateRole(Role request)
         {
             var role = _context.Roles.Find(request.Id);
+            if (role == null)
+                return false;
             role.Name = request.Name;
             role.Description = request.Description;
 
             _context.SaveChanges();
-            return _context.Roles.ToList();
+            return true;
         }
-        public List<Role> DeleteRole(Guid id)
+        public bool DeleteRole(Guid id)
         {
             var role = _context.Roles.Find(id);
-
+            if (role == null)
+                return false;
             _context.Roles.Remove(role);
             _context.SaveChanges();
-            return _context.Roles.ToList();
+            return true;
         }
     }
 }
